[IMP] mis_builder: move KpiMatrix impl to it's own file
